config/institutions/nyu-home.js: restore hardcoded scope [+]
`scope` was renamed `defaultScope` in https://github.com/NYULibraries/bess-vue/commit/f62b3fc6432134222c2d38b648b855184fe4abf8 (and `defaultScope` itself was moved to `nyu.js` later).  The redirect code drops the `scope` query param from the query string used for the `window.open` target URL because it's not defined.  Primo VE appears to default to the scope we want if `scope` query param is missing.  Adding it back for consistency with earlier version, and so we have finer control over the scope.